requirements.txt
Django==5.0
djangorestframework
python-docx
PyMuPDF
Pillow
pytesseract
python-dotenv
